--- a/engine/lite/k8sDeploymentManager/src/main/resources/defaultMinimalDeployment.conf
+++ b/engine/lite/k8sDeploymentManager/src/main/resources/defaultMinimalDeployment.conf
@@ -5,7 +5,7 @@
   },
   "spec": {
     # we set minReadySeconds default
-    "minReadySeconds": 10,
+    "minReadySeconds": 0,
     "selector": {
     },
     "template": {