--- a/engine/lite/k8sDeploymentManager/src/main/scala/pl/touk/nussknacker/k8s/manager/deployment/DeploymentPreparer.scala
+++ b/engine/lite/k8sDeploymentManager/src/main/scala/pl/touk/nussknacker/k8s/manager/deployment/DeploymentPreparer.scala
@@ -78,7 +78,7 @@ class DeploymentPreparer(config: K8sDeploymentManagerConfig) extends LazyLogging
 
   private def modifyContainers(containers: List[Container]): List[Container] = {
     val image = s"${config.dockerImageName}:${config.dockerImageTag}"
-    val defaultTimeoutSeconds = 3
+    val defaultTimeoutSeconds = 5
     val runtimeContainer = Container(
       name = "runtime",
       image = image,
@@ -97,14 +97,15 @@ class DeploymentPreparer(config: K8sDeploymentManagerConfig) extends LazyLogging
         Volume.Mount(name = "runtime-conf", mountPath = RuntimeConfigMountPath),
       ),
       // used standard AkkaManagement see HealthCheckServerRunner for details
-      // TODO we should tune failureThreshold to some lower value
-      readinessProbe = Some(Probe(new HTTPGetAction(Left(8080), path = "/ready"), periodSeconds = Some(1), failureThreshold = Some(60), timeoutSeconds = defaultTimeoutSeconds)),
+      startupProbe = Some(Probe(new HTTPGetAction(Left(8080), path = "/ready"), periodSeconds = Some(1), failureThreshold = Some(60), timeoutSeconds = defaultTimeoutSeconds)),
+      readinessProbe = Some(Probe(new HTTPGetAction(Left(8080), path = "/ready"), periodSeconds = Some(5), failureThreshold = Some(3), timeoutSeconds = defaultTimeoutSeconds)),
       livenessProbe = Some(Probe(new HTTPGetAction(Left(8080), path = "/alive"), periodSeconds = Some(5), failureThreshold = Some(3), timeoutSeconds = defaultTimeoutSeconds))
     )
 
     def modifyRuntimeContainer(value: Container): Container = {
       val containerLens = GenLens[Container](_.env).modify(_ ++ runtimeContainer.env) andThen
         GenLens[Container](_.volumeMounts).modify(_ ++ runtimeContainer.volumeMounts) andThen
+        GenLens[Container](_.startupProbe).modify(_.orElse(runtimeContainer.startupProbe)) andThen
         GenLens[Container](_.readinessProbe).modify(_.orElse(runtimeContainer.readinessProbe)) andThen
         GenLens[Container](_.livenessProbe).modify(_.orElse(runtimeContainer.livenessProbe)) andThen
         GenLens[Container](_.image).modify { configuredImage =>