- added startupProbe
- timeout seconds increased from default 1 sec to 5 secs (in case of e.g. full GC) 
 - liveness period: 10 sec -> 5 sec for faster unavailability detection 
 - minReadySeconds: 10 -> 0 (default value) to make runtime containers ready as soon as they are ready